pyproject.toml
[project]
name = "rag-pdf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "accelerate>=0.20.0",
    "aiohttp==3.9.1",
    "compel>=2.2.1",
    "diffusers>=0.27.0",
    "faiss-cpu==1.7.4",
    "fastapi==0.109.0",
    "huggingface-hub[hf-xet]>=0.35.3",
    "numpy==1.24.3",
    "peft>=0.17.1",
    "pillow>=10.0.0",
    "protobuf>=6.32.1",
    "pydantic==2.5.3",
    "python-dotenv==1.0.0",
    "python-multipart==0.0.6",
    "sentence-transformers==2.5.1",
    "sentencepiece>=0.2.1",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers==4.51.3",
    "adapters==1.2.0",
    "transformers>=4.40.0",
    "uvicorn[standard]==0.27.0",
    "llama-cpp-python>=0.3.16",
    "propcache==0.4.1",
    "adapter-transformers>=4.0.0",
]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu124" }
torchvision = { index = "pytorch-cu124" }

# diffusers = { git = "https://github.com/huggingface/diffusers.git" }
